spacing, refactoring
