+++ portal/src/components/FeatureCard.tsx
@@ -27,7 +27,7 @@ const FeatureCard = ({
         border="2px solid white"
         cursor="pointer"
         filter={disabled ? "grayscale(100%)" : ""}
-        minH="332px"
+        minH="386px"
       >
         <Image src={feature.image} w="170px" h="170px" alt={feature.name} mb="40px" />
         <Text fontSize="20px" fontWeight="500" mb="10px">
